agrego package.json y separo la logica del negocio de la persistencia. Llamo todo desde index.js
