+++ backend/Business/ProductManager.js
@@ -6,6 +6,8 @@ la cuál tendrá un arreglo de Productos que iniciará vacíos.
 
 */
 
+const Product = require('./Product');
+
 class ProductManager {
     constructor() {
         this.products = [];
@@ -33,7 +35,7 @@ class ProductManager {
         }
     
         const newId = this.getLastId() + 1;
-        const newProduct = { id: newId, title, description, price, thumbnail, stock };
+        const newProduct = new Product(newId, title, description, price, thumbnail, stock);
         this.products.push(newProduct);
         return newId;
     };
@@ -73,90 +75,4 @@ class ProductManager {
     };
 }
 
-let productAdmin = new ProductManager()
-try {
-    console.log("Prueba de getProducts(), debería retornar []");
-    console.log(productAdmin.getProducts())
-} catch (error) {
-    console.error(error.message);
-}
-
-console.log("\n -------------------");
-
-/* 
-    Agrega un producto utilizando el método addProduct
-*/
-
-
-const newProduct = {
-    title: "producto prueba",
-    description: "Este es un producto prueba",
-    price: 250,
-    thumbnail: "Sin imagen",
-    stock: 25
-};
-
-
-const prodAdmin = new ProductManager();
-try {
-    prodAdmin.addProduct(newProduct);
-    console.log(`Producto agregado con id: ${prodAdmin.getLastId()}`);
-} catch (error) {
-    console.error(error.message);
-}
-
-/*
-    Obtiene los productos actualizados con el método getProducts
-*/
-const products = prodAdmin.getProducts();
-
-console.log(products); // [{...}]
-
-
-/* Evaluar el método getProductById para obtener un producto por su id */
-
-const productId = 1; // ID del producto a buscar y mostrar
-
-
-try {
-    const product = prodAdmin.getProductById(productId);
-    console.log("Resultado de: getProductById(1):")
-    console.log(product); // { id: 1, ...}
-} catch (error) {
-    console.error(error.message);
-}
-
-console.log("\n -------------------");
-
-/* Editar un producto */
-
-// Editar un producto existente 
-
-const idProducto = 1; // ID del producto agregado anteriormente
-const newProductProps = {
-    price: 25.99,
-    description: "Esta es la nueva descripción del producto",
-};
-
-try {
-    const editedProduct = prodAdmin.editProduct(idProducto, newProductProps);
-    console.log("Resultado de: editProduct(idProducto, newProductProps)")
-    console.log(editedProduct);
-} catch (error) {
-    console.error(error.message);
-}
-
-console.log("\n -------------------");
-
-/*  Eliminar un producto */
-
-// Eliminar un producto existente
-const idProductoAEliminar = 1; // ID del producto agregado anteriormente
-
-try {
-    prodAdmin.deleteProduct(idProductoAEliminar);
-    console.log("Resultado de: deleteProduct(idProductoAEliminar);")
-    console.log(prodAdmin.getProducts())
-} catch (error) {
-    console.error(error.message);
-}
+module.exports = ProductManager;
